Incorporated feedback
Incorporate the feedback Alex gave us for the presentation.
Added quite some theory.

Unfortunately, not everything is grounded and referenced as I would like it to: From a scientific point of view it's good to justify what is written. in few cases, like why we add the GND plane, it's not, for a simple reason: I can't find an easy and reliable source to reference... Only internet blogs and stuff. And I am sorry, but I can't put more energy into this now. I feel like I have written my thesis. if something strikes bad, say it and we could remove it...